+++ shift-scan/src/app/(routes)/hamburger/inbox/_components/eqScannerModal.tsx
