+++ shift-scan/src/app/(routes)/hamburger/inbox/_components/viewDocumentContent.tsx
@@ -1,6 +1,5 @@
 "use client";
 import { Buttons } from "@/components/(reusable)/buttons";
-import { Contents } from "@/components/(reusable)/contents";
 import { Holds } from "@/components/(reusable)/holds";
 import { Titles } from "@/components/(reusable)/titles";
 import { useRouter } from "next/navigation";
